+++ packages/jest-haste-map/src/index.ts
@@ -13,7 +13,6 @@ import {EventEmitter} from 'events';
 import {tmpdir} from 'os';
 import * as path from 'path';
 import type {Stats} from 'graceful-fs';
-import {NodeWatcher, Watcher as SaneWatcher} from 'sane';
 import type {Config} from '@jest/types';
 import {escapePathForRegex} from 'jest-regex-util';
 import serializer from 'jest-serializer';
@@ -24,9 +23,6 @@ import H from './constants';
 import nodeCrawl = require('./crawlers/node');
 import watchmanCrawl = require('./crawlers/watchman');
 import getMockName from './getMockName';
-import FSEventsWatcher = require('./lib/FSEventsWatcher');
-// @ts-expect-error: not converted to TypeScript - it's a fork: https://github.com/facebook/jest/pull/5387
-import WatchmanWatcher from './lib/WatchmanWatcher';
 import * as fastPath from './lib/fast_path';
 import getPlatformExtension from './lib/getPlatformExtension';
 import normalizePathSep from './lib/normalizePathSep';
@@ -44,6 +40,11 @@ import type {
   ModuleMetaData,
   WorkerMetadata,
 } from './types';
+import FSEventsWatcher = require('./watchers/FSEventsWatcher');
+// @ts-expect-error: not converted to TypeScript - it's a fork: https://github.com/facebook/jest/pull/10919
+import NodeWatcher from './watchers/NodeWatcher';
+// @ts-expect-error: not converted to TypeScript - it's a fork: https://github.com/facebook/jest/pull/5387
+import WatchmanWatcher from './watchers/WatchmanWatcher';
 import {getSha1, worker} from './worker';
 // TypeScript doesn't like us importing from outside `rootDir`, but it doesn't
 // understand `require`.
@@ -97,7 +98,7 @@ type InternalOptions = {
 };
 
 type Watcher = {
-  close(callback: () => void): void;
+  close(): Promise<void>;
 };
 
 type WorkerInterface = {worker: typeof worker; getSha1: typeof getSha1};
@@ -210,7 +211,7 @@ function invariant(condition: unknown, message?: string): asserts condition {
 export default class HasteMap extends EventEmitter {
   private _buildPromise: Promise<InternalHasteMapObject> | null;
   private _cachePath: Config.Path;
-  private _changeInterval?: NodeJS.Timeout;
+  private _changeInterval?: ReturnType<typeof setInterval>;
   private _console: Console;
   private _options: InternalOptions;
   private _watchers: Array<Watcher>;
@@ -253,14 +254,8 @@ export default class HasteMap extends EventEmitter {
           options.ignorePattern.flags,
         );
       } else {
-        const ignorePattern = options.ignorePattern;
-        const vcsIgnoreRegExp = new RegExp(VCS_DIRECTORIES);
-        this._options.ignorePattern = (filePath: string) =>
-          vcsIgnoreRegExp.test(filePath) || ignorePattern(filePath);
-
-        this._console.warn(
-          'jest-haste-map: the `ignorePattern` options as a function is being ' +
-            'deprecated. Provide a RegExp instead. See https://github.com/facebook/jest/pull/4063.',
+        throw new Error(
+          'jest-haste-map: the `ignorePattern` option must be a RegExp',
         );
       }
     } else {
@@ -781,7 +776,7 @@ export default class HasteMap extends EventEmitter {
     this._options.retainAllFiles = true;
 
     // WatchmanWatcher > FSEventsWatcher > sane.NodeWatcher
-    const Watcher: SaneWatcher =
+    const Watcher =
       canUseWatchman && this._options.useWatchman
         ? WatchmanWatcher
         : FSEventsWatcher.isSupported()
@@ -798,7 +793,6 @@ export default class HasteMap extends EventEmitter {
     let mustCopy = true;
 
     const createWatcher = (root: Config.Path): Promise<Watcher> => {
-      // @ts-expect-error: TODO how? "Cannot use 'new' with an expression whose type lacks a call or construct signature."
       const watcher = new Watcher(root, {
         dot: true,
         glob: extensions.map(extension => '**/*.' + extension),
@@ -824,10 +818,7 @@ export default class HasteMap extends EventEmitter {
         mustCopy = true;
         const changeEvent: ChangeEvent = {
           eventsQueue,
-          hasteFS: new HasteFS({
-            files: hasteMap.files,
-            rootDir,
-          }),
+          hasteFS: new HasteFS({files: hasteMap.files, rootDir}),
           moduleMap: new HasteModuleMap({
             duplicates: hasteMap.duplicates,
             map: hasteMap.map,
@@ -1043,20 +1034,18 @@ export default class HasteMap extends EventEmitter {
     }
   }
 
-  end(): Promise<void> {
-    // @ts-expect-error: TODO TS cannot decide if `setInterval` and `clearInterval` comes from NodeJS or the DOM
-    clearInterval(this._changeInterval);
-    if (!this._watchers.length) {
-      return Promise.resolve();
+  async end(): Promise<void> {
+    if (this._changeInterval) {
+      clearInterval(this._changeInterval);
     }
 
-    return Promise.all(
-      this._watchers.map(
-        watcher => new Promise(resolve => watcher.close(resolve)),
-      ),
-    ).then(() => {
-      this._watchers = [];
-    });
+    if (!this._watchers.length) {
+      return;
+    }
+
+    await Promise.all(this._watchers.map(watcher => watcher.close()));
+
+    this._watchers = [];
   }
 
   /**