+++ packages/jest-types/src/Config.ts
@@ -64,6 +64,7 @@ export type DefaultOptions = {
   coverageProvider: CoverageProvider;
   errorOnDeprecated: boolean;
   expand: boolean;
+  extensionsToTreatAsEsm: Array<Path>;
   forceCoverageMatch: Array<Glob>;
   globals: ConfigGlobals;
   haste: HasteConfig;
@@ -138,6 +139,7 @@ export type InitialOptions = Partial<{
   detectOpenHandles: boolean;
   displayName: string | DisplayName;
   expand: boolean;
+  extensionsToTreatAsEsm: Array<Path>;
   extraGlobals: Array<string>;
   filter: Path;
   findRelatedTests: boolean;
@@ -323,6 +325,7 @@ export type ProjectConfig = {
   detectOpenHandles: boolean;
   displayName?: DisplayName;
   errorOnDeprecated: boolean;
+  extensionsToTreatAsEsm: Array<Path>;
   extraGlobals: Array<keyof NodeJS.Global>;
   filter?: Path;
   forceCoverageMatch: Array<Glob>;